Fix navigation
This is hacky and probably not at all the way to do this, but it does
work. We use vue-router, which allows us to modify the URL quite nicely
and pass values from the querystring to our new 'Board' component. This
requires a server to function correctly, so I've documented that in the
README.

Note that this likely breaks deployment of this page for anyone but me
deploying on the local server. There are two reason for this. Firstly,
I'm hardcoding the root URL to match that of a local server I'm using
that has the Apache 'UserDir' directive [1] enabled. Secondly, you're
going to see CORS issues if you're not deploying on a '*.redhat.com'
server because 'vue-router' requires a server rather than just the
'index.html' file, but the use of a server triggers the browser's CORS
checks [2].

[1] https://httpd.apache.org/docs/2.4/howto/public_html.html
[2] https://developer.mozilla.org/en-US/docs/Web/HTTP/CORS

Signed-off-by: Stephen Finucane <stephen@that.guru>